use unsigned int over size_t

+++ src/fmi4cpp/fmi2/xml/ModelDescriptionParser.hpp
@@ -144,7 +144,7 @@ namespace {
     CoSimulationAttributes parseCoSimulationAttributes(const ptree &node) {
 
         auto commonAttributes = parseFmuAttributes(node);
-        auto maxOutputDerivativeOrder = node.get<size_t>("<xmlattr>.maxOutputDerivativeOrder", 0);
+        auto maxOutputDerivativeOrder = node.get<unsigned int>("<xmlattr>.maxOutputDerivativeOrder", 0);
         auto canInterpolateInputs = node.get<bool>("<xmlattr>.canInterpolateInputs", false);
         auto canRunAsynchronuously = node.get<bool>("<xmlattr>.canRunAsynchronuously", false);
         auto canHandleVariableCommunicationStepSize = node.get<bool>("<xmlattr>.canHandleVariableCommunicationStepSize",
